Second step: Getting the list of pokemons
It is horrible to see an asyncronous call inside another one, I know.
We'll deal with that later. At least we have the second list too. The
structure of this API is quite more complex. I'm extracting the data
"the easy way", but it is not the correct way though. We should access
to the specific fields from inside each specific page. If there's some
time left, we'll correct this later.