Third Step: Getting the last list and concatenating
This end point is weird...because there is no ID per se. We have the
'driverId'... but in that case, should I use the 'nameId' in the card
list?? this is confusing. Anyway, third list ready, and all
concatenated. Next step: return the result to the browser when the
content is requested.